Update patch set 4
Patch Set 4:

Hi,

Renaming the folder is something not needed, and it should be better if you keep it as is, as this is the way Ansible 2.0 handle the vars. I asumme compass will go for this version at some point.

About the packages, if we guarantee they are in place while building, it's fine to drop them. I'll revise the package stuff in case I'm in time to do it.

Thanks!

Jos

Patch-set: 4
Label: Verified=0