Update the API version for Openstack Newton
Use the "keystone-manage bootstrap" command to instead of admin_token.
Because the  admin_token is treated as a "shared secret" that can be used
to bootstrap Keystone through the API. This "token" does not represent a
user (it has no identity), and carries no explicit authorization (it
effectively bypasses most authorization checks).

Use the API v3 to instead of API v2.0. Identity API v3 was established to
introduce namespacing for users and projects by using "domains" as a
higher-level container for more flexible identity management and fixed a
security issue in the v2.0 API (bearer tokens appearing in URLs).

JIRA: COMPASS-491

Change-Id: I56182c14b761728c3492b9dd2b05c3b57aa5f35f
Signed-off-by: liyuenan <liyuenan@huawei.com>