FIX external net create
JIRA: -

The dictionary public_net_info in network.yml had changed to pass the
yamllint:
  public_net_info:
    enable: True
change to:
  public_net_info:
    enable: "True"
Because of that, the public_net_info.enable in group_var/all had changed.
  'public_net_info': {..., 'enable': True, ...}
change to:
  'public_net_info': {..., 'enable': 'True', ...}
So the conditionals in roles/ext-network/tasks/main.yml should be changed,
otherwise the ext-net can not be created.

Change-Id: I7b4aeb1d0edec0ad77252849704bdf3055adfa5a
Signed-off-by: liyuenan <liyuenan@huawei.com>