deploy/conf: separate x86 and AArch64 docker image naming
Latest update hard-coded some docker image names, which breaks
the deployment on Arm platform. This patch fixes the issue by
distinguishing docker image naming on different architecture.

This is an interim solution, and in the future, we may leverage
docker multi arch support to use consistent naming for
different architectures.

Change-Id: I5c59c12fb250c4e09d83582640e248075b5f7373
Signed-off-by: Yibo Cai <yibo.cai@arm.com>