Enable neutron port trunk
JIRA: COMPASS-571

The network trunk service allows multiple networks to be connected to an
instance using a single virtual NIC (vNIC). Multiple networks can be presented
to an instance by connecting it to a single port. This is a highly relevant
feature for VNFs and hence for OPNFV as a reference NFVI.

This patch is about enabling the trunk port feature in Compass deployments.

Change-Id: I270c3f1d5e4af99578c8e1882fae12ddddaabe78
Signed-off-by: Yifei Xue <xueyifei@huawei.com>