committer nomination for Harry Huang
Hi team, I propose to nominate Harry as a committer of Compass4nfv.
Harry did a lots of work in containized compass core and he is so kind to
help fix Compass4nfv issues in Fraser release.

More detailed contributions are listed:
https://gerrit.opnfv.org/gerrit/#/q/owner:huangxiangyu5%2540huawei.com+status:merged+projects:compass4nfv

please vote +2 if you agree with Harry's nomination.

Change-Id: I3a8cbbe649688cb60271375bbfc09aa0eadfae3b
Signed-off-by: chigang <chigang@huawei.com>