Workaround for Tacker in HA scenarios
JIRA: -

For Tacker service in HA mode, it only generates VIM fernet key on
one of the Tacker hosts. This may cause error when a request is
been routed to another host. We need to use Barbican and Mistral
components. But now OSA did not support Mistral yet. So we add
a workaround here to get over this issue.

Change-Id: I9aa228ec2ac7011cedb1eb503611ca24c7666caa
Signed-off-by: Yifei Xue <xueyifei@huawei.com>