Change the dashboard and heapster images in k8s to arm64 when running on arm platform
Change the dashboard image in k8s from amd64 to arm64 on arm platform.
In extra-vars-aarch64.yml, add the dashboard image tag for supporting the
arm64 arch.

For heapster images, add a "sed" command in main.yml for replacing
the cpu architecture from amd64 to arm64.
And it will be invoked only when deploying the arm platform.

Change-Id: I584485f0356dcb925814f63cbc45ecbd459715bb
Signed-off-by: Jingzhao Ni <jingzhao.ni@arm.com>